refactor(commands): update import paths for simpleEmbed and remove unused imports
refactor(embed): move embed utility function to a new file and rename it to discordEmbed to improve code organization and semantics
