+++ src/commands/anniversaire/[sub-commands]/list.cmd.ts
@@ -2,7 +2,7 @@ import type { CommandExecute } from "$core/utils/handler/command";
 import { getBirthdays } from "$core/api/requests/member";
 import { memberPerPage } from "./list.const";
 import { DayJS } from "$core/configs/day-js";
-import { simpleEmbed } from "$core/utils/embed";
+import { simpleEmbed } from "$core/utils/discord/embed";
 import { dateFormat, getAge } from "$core/utils/function";
 import { msgParams } from "$core/utils/message";
 import { gqlRequest } from "$core/utils/request";
