+++ src/commands/forum/[sub-commands]/why.cmd.ts
@@ -1,4 +1,4 @@
-import { simpleEmbed } from "$core/utils/embed";
+import { simpleEmbed } from "$core/utils/discord/embed";
 import type { CommandExecute } from "$core/utils/handler/command";
 import { commands } from "$core/configs/message/command";
 import { msgParams } from "$core/utils/message";
