+++ src/commands/interaction/[sub-commands]/verif.cmd.ts
@@ -1,5 +1,5 @@
 import { candidatBtn } from "./verif.util";
-import { simpleEmbed } from "$core/utils/embed";
+import { simpleEmbed } from "$core/utils/discord/embed";
 import type { CommandExecute } from "$core/utils/handler/command";
 import { commands } from "$core/configs/message/command";
 import { ChannelType } from "discord.js";
