+++ src/commands/main/[sub-commands]/group-add/add-role.cmd.ts
@@ -1,6 +1,6 @@
 import type { CommandExecute } from "$core/utils/handler/command";
 import { addRole } from "$core/api/requests/main-role";
-import { simpleEmbed } from "$core/utils/embed";
+import { simpleEmbed } from "$core/utils/discord/embed";
 import { commands } from "$core/configs/message/command";
 import { gqlRequest } from "$core/utils/request";
 import { msgParams } from "$core/utils/message";
