+++ src/commands/main/[sub-commands]/group-remove/remove-role.cmd.ts
@@ -1,6 +1,6 @@
 import type { CommandExecute } from "$core/utils/handler/command";
 import { removeRole } from "$core/api/requests/main-role";
-import { simpleEmbed } from "$core/utils/embed";
+import { simpleEmbed } from "$core/utils/discord/embed";
 import { msgParams } from "$core/utils/message";
 import { gqlRequest } from "$core/utils/request";
 import { commands } from "$core/configs/message/command";
