+++ src/events/roles-selector/roles-selector.event.ts
@@ -1,8 +1,8 @@
 import type { EventExecute, EventName } from "$core/utils/handler/event";
-import { interactionId } from "$core/configs";
+import { interactionId } from "$core/configs/global";
 import { guilds } from "$core/configs/guild";
 import { events } from "$core/configs/message/event";
-import { simpleEmbed } from "$core/utils/embed";
+import { simpleEmbed } from "$core/utils/discord/embed";
 import { logger } from "$core/utils/logger";
 import { msgParams } from "$core/utils/message";
 import { GuildMemberRoleManager } from "discord.js";
