+++ src/tasks/birthday-checker/birthday-checker.task.ts
@@ -4,7 +4,7 @@ import { client } from "$core/client";
 import { getGuild, guilds } from "$core/configs/guild";
 import { tasks } from "$core/configs/message/task/task.config";
 import { DayJS } from "$core/configs/day-js";
-import { simpleEmbed } from "$core/utils/embed";
+import { simpleEmbed } from "$core/utils/discord/embed";
 import { getAge } from "$core/utils/function";
 import { logger } from "$core/utils/logger";
 import { msgParams } from "$core/utils/message";
